+++ src/XRP/xrp-client-decorator.ts
@@ -2,7 +2,7 @@ import { Wallet } from 'xpring-common-js'
 import { BigInteger } from 'big-integer'
 import TransactionStatus from './transaction-status'
 import RawTransactionStatus from './raw-transaction-status'
-import XRPTransaction from './model/xrp-transaction'
+import { XRPTransaction } from './model/xrp-transaction'
 import { XRPLNetwork } from '../Common/xrpl-network'
 import SendXrpDetails from './model/send-xrp-details'