+++ test/XRP/fakes/fake-xrp-transaction-type-protobufs.ts
@@ -1,8 +1,12 @@
+import {
+  AccountSet,
+  AccountDelete,
+  CheckCancel,
+} from '../../../src/XRP/Generated/web/org/xrpl/rpc/v1/transaction_pb'
 /* eslint-disable @typescript-eslint/no-magic-numbers --
  * ESLint flags the numbers in the Uint8Array as magic numbers,
  * but this is a fakes file for testing, so it's fine.
  */
-
 import {
   ClearFlag,
   Domain,
@@ -11,12 +15,16 @@ import {
   SetFlag,
   TransferRate,
   TickSize,
+  Destination,
+  DestinationTag,
+  CheckID,
 } from '../../../src/XRP/Generated/web/org/xrpl/rpc/v1/common_pb'
-import { AccountSet } from '../../../src/XRP/Generated/web/org/xrpl/rpc/v1/transaction_pb'
+import { AccountAddress } from '../../../src/XRP/Generated/web/org/xrpl/rpc/v1/account_pb'
 
 // Primitive test values ===============================================================
 
-// AccountSet
+// AccountSet values
+
 const testClearFlag = 5
 const testDomain = 'testdomain'
 const testEmailHash = new Uint8Array([8, 9, 10])
@@ -24,6 +32,14 @@ const testMessageKey = new Uint8Array([11, 12, 13])
 const testSetFlag = 4
 const testTransferRate = 1234567890
 const testTickSize = 7
+
+// AccountDelete values
+const testDestination = 'rPEPPER7kfTD9w2To4CQk6UCfuHM9c6GDY'
+const testDestinationTag = 13
+
+// CheckCancel values
+const testCheckId =
+  '49647F0D748DC3FE26BDACBC57F251AADEFFF391403EC9BF87C97F67E9977FB0'
 
 // Protobuf objects ======================================================================
 
@@ -62,4 +78,39 @@ testAccountSetProtoAllFields.setTickSize(testTickSizeProto)
 const testAccountSetProtoOneFieldSet = new AccountSet()
 testAccountSetProtoOneFieldSet.setClearFlag(testClearFlagProto)
 
-export { testAccountSetProtoAllFields, testAccountSetProtoOneFieldSet }
+// AccountDelete protos
+const testAccountAddressProto = new AccountAddress()
+testAccountAddressProto.setAddress(testDestination)
+
+const testDestinationProto = new Destination()
+testDestinationProto.setValue(testAccountAddressProto)
+
+const testDestinationTagProto = new DestinationTag()
+testDestinationTagProto.setValue(testDestinationTag)
+
+const testAccountDeleteProto = new AccountDelete()
+testAccountDeleteProto.setDestination(testDestinationProto)
+testAccountDeleteProto.setDestinationTag(testDestinationTagProto)
+
+const testAccountDeleteProtoNoTag = new AccountDelete()
+testAccountDeleteProtoNoTag.setDestination(testDestinationProto)
+
+// CheckCancel proto
+const testCheckIdProto = new CheckID()
+testCheckIdProto.setValue(testCheckId)
+
+const testCheckCancelProto = new CheckCancel()
+testCheckCancelProto.setCheckId(testCheckIdProto)
+
+// Invalid Protobuf Objects ========================================================================
+
+const testInvalidCheckCancelProto = new CheckCancel()
+
+export {
+  testAccountSetProtoAllFields,
+  testAccountSetProtoOneFieldSet,
+  testAccountDeleteProto,
+  testAccountDeleteProtoNoTag,
+  testCheckCancelProto,
+  testInvalidCheckCancelProto,
+}